voxcode/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voxcode"
version = "0.1.0"
description = "Open-source voice coding assistant with AI superpowers"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "VoxCode Contributors", email = "voxcode@lordburnitdown.dev"}
]
keywords = ["voice", "coding", "assistant", "whisper", "ai", "ide"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openai-whisper>=20231117",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "pynput>=1.7.6",
    "pyperclip>=1.8.2",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "watchdog>=3.0.0",
    "platformdirs>=4.0.0",
    "anthropic>=0.45.0",
]

[project.optional-dependencies]
gpu = ["torch>=2.0.0"]
ui = ["pystray>=0.19.0", "pillow>=10.0.0"]
code-intel = ["tree-sitter>=0.20.0", "tree-sitter-languages>=1.8.0"]
server = ["fastapi>=0.104.0", "uvicorn>=0.24.0", "websockets>=12.0"]
all = ["voxcode[gpu,ui,code-intel,server]"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
voxcode = "voxcode.__main__:main"

[project.urls]
Homepage = "https://github.com/LORDBurnitdown/voxcode"
Documentation = "https://github.com/LORDBurnitdown/voxcode#readme"
Repository = "https://github.com/LORDBurnitdown/voxcode"
Issues = "https://github.com/LORDBurnitdown/voxcode/issues"
Changelog = "https://github.com/LORDBurnitdown/voxcode/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
